backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18
python-multipart==0.0.6
python-dotenv==1.0.0
pdfplumber==0.10.3
python-docx==1.0.1
pydantic==2.5.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiofiles==23.2.1 